refactor(workspace): 5-phase audit remediation -- edition 2024, dep centralization, dedup
Phase 1 (P0): Centralize dependency versions
- rustfft, faer, proptest, burn/blas-src/openblas-src -> workspace = true
- pgfplots, chrono, latex, approx -> workspace = true
- Fix pre-existing API drift in lbm_3d_cuda bench files (Precision arg, f32 types)

Phase 2 (P1): Normalize workspace manifests
- Migrate all 26 crates from edition 2021 to edition 2024 (workspace-inherited)
- Add resolver = "3" (edition 2024 default)
- Add [workspace.lints] with clippy::all warn + rust warnings deny
- Normalize dep.workspace = true -> dep = { workspace = true } (8 crates)
- Fix edition 2024 breaking changes:
  - gen keyword reservation: 29 rng.gen() -> rng.r#gen(), 6 variable renames
  - Pattern binding modes: 17 filter/find closures with implicit borrow conflicts
  - unsafe set_var: 9 std::env::set_var() calls wrapped in unsafe blocks
  - collapsible_if: 29+ if-let chains collapsed (new edition 2024 feature)
  - ref in implicit borrow: 1 occurrence simplified

Phase 3 (P2): Code deduplication
- Extract canonical parse_f64_or_nan/parse_f64_or_zero to data_core::parse
- Extract canonical mean/std_dev/median to stats_core::helpers
- Replace 12 private parse_f64 implementations with canonical import
- Replace 4 private stats helper implementations with canonical import
- Net ~250 lines removed

Phase 4 (P1): Registry alignment
- Register 6 missing binaries in binaries.toml
- Update project.toml counters (claims 700, experiments 36, tests 3524)
- Update agents.toml (workspace_crates=26, cli_binaries=101)

Phase 5 (P3): Governance reconciliation
- Update MEMORY.md with current counts, edition 2024 pitfalls, new modules
- Final gate: 0 clippy warnings, 3524 tests passing, edition 2024 workspace-wide

137 files changed, 754 insertions(+), 842 deletions(-) (net -88 lines)

+++ crates/cosmic_scheduler/tests/test_lbm_integration.rs
@@ -53,7 +53,7 @@ impl TwoPhaseSystem for MockLbmSystem {
             // Simple velocity averaging (simplified from D3Q19)
             if self.rho[i] > 0.0 {
                 self.u[i][0] =
-                    (self.f[i * 19 + 1] as f64 - self.f[i * 19 + 2] as f64) / self.rho[i];
+                    (self.f[i * 19 + 1] - self.f[i * 19 + 2]) / self.rho[i];
             }
         }
